// dfp-piao-fe/src/app/shared/models/enums/piao-status.enum.ts
export enum PiaoStatusEnum {
  DA_COMPILARE = 'Da compilare',
  IN_COMPILAZIONE = 'In compilazione',
  COMPILATO = 'Compilato',
  IN_VALIDAZIONE = 'In validazione',
  VALIDATO = 'Validato',
  RICHIESTA_APPROVAZIONE = 'Richiesta Approvazione',
  APPROVATO = 'Approvato',
  PUBBLICATO = 'Pubblicata',
}